Release 1.0.0 (#17)
* feat: spring security setup

This allows user authorization with jwt

* feat: exceptions

This allows custom exception throwing.

* User service & controller (#3)

* feat: organize user service (WIP)

This sets up a basic plan for this microservice. It organises entities, services, controllers, etc.

* feat: login and refactor (WIP)

This allows users to login to the application.

* feat: user service and controller crud

This allows users to create accounts, login, edit their accounts, change passwords, etc.

* Auth refactor (#4)

* feat: organize user service (WIP)

This sets up a basic plan for this microservice. It organises entities, services, controllers, etc.

* feat: login and refactor (WIP)

This allows users to login to the application.

* feat: user service and controller crud

This allows users to create accounts, login, edit their accounts, change passwords, etc.

* feat: user service and controller refactor

This moves register endpoint to auth controller instead of user controller.

* feat: user service and controller refactor

* feat: properties and gradle setup + refactored auth (#5)

This sets up application.properties and fixes dependencies. Code is refactored so that auth works fine.

* Tests setup (#6)

* feat: test setup and user service integration tests

This sets up integration tests. UserService integration tests are written.

* feat: service and controller integration tests

* feat: get user service and controller integration tests

* feat: unit tests

* feat: user delete endpoint (#7)

* fix: tests container db config (#8)

* feat: docker and docker compose (#9)

This sets up a Dockerfile and also docker-compose.yml with database for this service. Properties file is changed to point to database container

* Docker build fix (#11)

* feat: docker and docker compose

This sets up a Dockerfile and also docker-compose.yml with database for this service. Properties file is changed to point to database container

* fix: docker build with tests

* feat: user notifications enable/disable (#12)

* feat: docker network setup (#13)

* fix: postgres and docker compose refactored, docker network fix (#14)

* fix: datasource url (#15)

* CI/CD (#16)

* feat: ci/cd

* feat: ci/cd integration tests

* feat: ci/cd publish to dockerhub

* feat: publish to dockerhub refactor

---------

Co-authored-by: Mihajlo Djordjevic <75526571+mihajlodj@users.noreply.github.com>